requirements.txt
pandas
scikit-learn
fastapi
uvicorn
seaborn
boto3
python-dotenv